IT: extract CapturingRetryPolicy to utils.rs
This util will be used by another regression test in session.rs in the
next commit.
